# Vectorize reduction when reducing on fastest striding dimension