Stack from [ghstack](https://github.com/ezyang/ghstack):
* **#36709 Vectorize reduction when reducing on fastest striding dimension**

Differential Revision: [D21083393](https://our.internmc.facebook.com/intern/diff/D21083393)